DjangoLearn/requirements.txt
djangorestframework==3.6.3
django-cors-headers==2.1.0